+++ apps/webapp/app/presenters/v3/EditSchedulePresenter.server.ts
@@ -88,6 +88,7 @@ export class EditSchedulePresenter {
     const schedule = await this.#prismaClient.taskSchedule.findFirst({
       select: {
         id: true,
+        type: true,
         friendlyId: true,
         generatorExpression: true,
         externalId: true,