+++ apps/webapp/app/v3/services/replayTaskRun.server.ts
@@ -1,9 +1,10 @@
-import { conditionallyImportPacket, parsePacket } from "@trigger.dev/core/v3";
+import { conditionallyImportPacket, parsePacket, RunTags } from "@trigger.dev/core/v3";
 import { TaskRun } from "@trigger.dev/database";
 import { findEnvironmentById } from "~/models/runtimeEnvironment.server";
 import { logger } from "~/services/logger.server";
 import { BaseService } from "./baseService.server";
 import { OutOfEntitlementError, TriggerTaskService } from "./triggerTask.server";
+import { getTagsForRunId } from "~/models/taskRunTag.server";
 
 export class ReplayTaskRunService extends BaseService {
   public async call(existingTaskRun: TaskRun) {
@@ -36,6 +37,11 @@ export class ReplayTaskRunService extends BaseService {
     });
 
     try {
+      const tags = await getTagsForRunId({
+        friendlyId: existingTaskRun.id,
+        environmentId: authenticatedEnvironment.id,
+      });
+
       const triggerTaskService = new TriggerTaskService();
       return await triggerTaskService.call(
         existingTaskRun.taskIdentifier,
@@ -49,6 +55,7 @@ export class ReplayTaskRunService extends BaseService {
             concurrencyKey: existingTaskRun.concurrencyKey ?? undefined,
             test: existingTaskRun.isTest,
             payloadType: payloadPacket.dataType,
+            tags: tags?.map((t) => t.name) as RunTags,
           },
         },
         {