+++ packages/database/prisma/migrations/20240717101035_add_task_schedule_type_dynamic_and_static/migration.sql
@@ -1,0 +1,6 @@
+-- CreateEnum
+CREATE TYPE "ScheduleType" AS ENUM ('DECLARATIVE', 'IMPERATIVE');
+
+-- AlterTable
+ALTER TABLE "TaskSchedule"
+ADD COLUMN "type" "ScheduleType" NOT NULL DEFAULT 'IMPERATIVE';