+++ integrations/replicate/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @trigger.dev/replicate
 
+## 3.0.0-beta.51
+
+### Patch Changes
+
+- Updated dependencies [979bee50d]
+- Updated dependencies [086a0f95c]
+- Updated dependencies [55264657d]
+  - @trigger.dev/sdk@3.0.0-beta.51
+  - @trigger.dev/integration-kit@3.0.0-beta.51
+
+## 3.0.0-beta.50
+
+### Patch Changes
+
+- Updated dependencies [8ba998794]
+  - @trigger.dev/sdk@3.0.0-beta.50
+  - @trigger.dev/integration-kit@3.0.0-beta.50
+
 ## 3.0.0-beta.49
 
 ### Patch Changes