+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,21 @@
 # create-trigger
 
+## 3.0.0-beta.51
+
+### Patch Changes
+
+- Updated dependencies [55264657d]
+  - @trigger.dev/core@3.0.0-beta.51
+  - @trigger.dev/yalt@3.0.0-beta.51
+
+## 3.0.0-beta.50
+
+### Patch Changes
+
+- Updated dependencies [8ba998794]
+  - @trigger.dev/core@3.0.0-beta.50
+  - @trigger.dev/yalt@3.0.0-beta.50
+
 ## 3.0.0-beta.49
 
 ### Patch Changes