+++ packages/eslint-plugin/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @trigger.dev/eslint-plugin
 
+## 3.0.0-beta.51
+
+## 3.0.0-beta.50
+
 ## 3.0.0-beta.49
 
 ## 3.0.0-beta.48